+++ CHANGELOG.md
@@ -8,6 +8,7 @@ _Unreleased_
 * (+) buttons for the first two modes can also be dragged onto the map.
 * Removing preset tags properly.
 * Turns out "Cancel" buttons for buildings and entrances kept the changes.
+* Correct weekday order and default holidays for various countries.
 * Added `sr-Latn` language for Serbia (thanks @alnzrv).
 * Updated roof assets (thanks @mxdanger).
 * Updated a phone number parsing library.